Fix min/max value, add disallowed values for age, refactor
